pyproject.toml
[tool.poetry]
name = "jim-emacs-fun-py"
version = "0.1.0"
description = ""
authors = ["stevechan <chanshunli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
llama-index-core = "^0.10.57"
llama-index-llms-openai = "^0.1.26"
llama-index-llms-replicate = "^0.1.3"
llama-index-embeddings-huggingface = "^0.2.2"
python-dotenv = "^1.0.1"
llama-index-llms-ollama = "^0.1.6"
llama-index-readers-file = "^0.1.30"
llama-index-embeddings-openai = "^0.1.11"
llama-parse = "^0.4.9"
llama-index-tools-google = "^0.1.6"
llama-index-tools-yahoo-finance = "^0.1.1"
flask-socketio = "^5.3.6"
flask = "^3.0.3"

[tool.poetry.group.dev.dependencies]
langfuse = "^2.36.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
